+++ changelog.md
@@ -55,3 +55,4 @@
 ## 0.3.10
 - Added a Team logo for the label
 - Ability to change logo on label
+- Handle cases where logo is null